block-builder-scheduler: move offset probing into its own file
Relocate the offset-probing code (offsetTime, probeInitialOffsets,
offsetStore, offsetFinder) and its tests out of scheduler.go/scheduler_test.go
into offset_scanner.go/offset_scanner_test.go. Pure code movement, no
behavior change.
